+++ indispensable-asn1/src/commonMain/kotlin/at/asitplus/signum/indispensable/asn1/Asn1Elements.kt
@@ -50,23 +50,27 @@ sealed class Asn1Element(
     }
 
     /**
-     * Length (already properly encoded into a byte array for writing as ASN.1) of the contained data.
+     * Length (already properly encoded into a byte array for writing as ASN.1) **of the contained data**, not the whole
+     * ASN.1 element.
      * For a primitive, this is just the size of the held bytes.
      * For a structure, it is the sum of the number of bytes needed to encode all held child nodes.
      */
-    val encodedLength by lazy { length.encodeLength() }
+    val encodedLength by lazy { contentLength.encodeLength() }
 
     /**
      * Length (as a plain `Int` to work with it in code) of the contained data.
      * For a primitive, this is just the size of the held bytes.
      * For a structure, it is the sum of the number of bytes needed to encode all held child nodes.
      */
-    abstract val length: Int
+    abstract val contentLength: Int
+
+    @Deprecated("unclear name", ReplaceWith("contentLength"))
+    val length get() = contentLength
 
     /**
      * Total number of bytes required to represent the ths element, when encoding to ASN.1.
      */
-    val overallLength by lazy { length + tag.encodedTagLength + encodedLength.size }
+    val overallLength by lazy { contentLength + tag.encodedTagLength + encodedLength.size }
 
 
     private val derEncodedLazy = lazy { Buffer().also { encodeTo(it) }.readByteArray() }
@@ -96,7 +100,7 @@ sealed class Asn1Element(
     fun prettyPrint() = prettyPrint(0)
 
     protected open fun prettyPrintHeader(indent: Int) =
-        "(tag=${tag}" + ", length=${length}" + ", overallLength=${overallLength})"
+        "(tag=${tag}" + ", length=${contentLength}" + ", overallLength=${overallLength})"
 
     protected open fun prettyPrintTrailer(indent: Int) = ""
     protected abstract fun prettyPrintContents(indent: Int): String
@@ -182,7 +186,7 @@ sealed class Asn1Element(
 
     @Throws(Asn1StructuralException::class)
     private inline fun <reified T> thisAs(): T =
-       (this as? T)
+        (this as? T)
             ?: throw Asn1StructuralException("${this::class.simpleName} cannot be reinterpreted as ${T::class.simpleName}.")
 
 
@@ -311,6 +315,31 @@ sealed class Asn1Element(
             val TIME_GENERALIZED = Tag(tagValue = BERTags.GENERALIZED_TIME.toULong(), constructed = false)
             val TIME_UTC = Tag(tagValue = BERTags.UTC_TIME.toULong(), constructed = false)
 
+            val entries: Iterable<Tag> by lazy {
+                listOf(
+                    SET,
+                    SEQUENCE,
+                    NULL,
+                    BOOL,
+                    INT,
+                    REAL,
+                    OID,
+                    ENUM,
+                    OCTET_STRING,
+                    BIT_STRING,
+                    STRING_UTF8,
+                    STRING_UNIVERSAL,
+                    STRING_IA5,
+                    STRING_BMP,
+                    STRING_T61,
+                    STRING_PRINTABLE,
+                    STRING_NUMERIC,
+                    STRING_VISIBLE,
+                    TIME_GENERALIZED,
+                    TIME_UTC
+                )
+            }
+
         }
 
         val tagClass by lazy {
@@ -319,13 +348,39 @@ sealed class Asn1Element(
             }
         }
 
+        val name
+            get() = when (this) {
+                SET -> "SET"
+                SEQUENCE -> "SEQUENCE"
+                NULL -> "NULL"
+                BOOL -> "BOOLEAN"
+                INT -> "INTEGER"
+                REAL -> "REAL"
+                OID -> "OBJECT IDENTIFIER"
+                ENUM -> "ENUMERATED"
+                OCTET_STRING -> "OCTET STRING"
+                BIT_STRING -> "BIT STRING"
+                STRING_UTF8 -> "UTF8 STRING"
+                STRING_UNIVERSAL -> "UNIVERSAL STRING"
+                STRING_IA5 -> "IA5 STRING"
+                STRING_BMP -> "BMP STRING"
+                STRING_T61 -> "T61 STRING"
+                STRING_PRINTABLE -> "PRINTABLE STRING"
+                STRING_NUMERIC -> "NUMERIC STRING"
+                STRING_VISIBLE -> "VISIBLE STRING"
+                TIME_GENERALIZED -> "GENERALIZED TIME"
+                TIME_UTC -> "UTC TIME"
+                else -> null
+            }
+
+
         val isConstructed get() = encodedTag.first().toUByte().isConstructed()
 
         internal val isExplicitlyTagged get() = isConstructed && tagClass == TagClass.CONTEXT_SPECIFIC
 
         override fun toString(): String =
             "${tagClass.let { if (it == TagClass.UNIVERSAL) "" else it.name + " " }}${tagValue}${if (isConstructed) " CONSTRUCTED" else ""}" +
-                    (" (=${encodedTag.toHexString(HexFormat.UpperCase)})")
+                    (" (=${encodedTag.toHexString(HexFormat.UpperCase)})" + (name?.let { " ($it)" } ?: ""))
 
         /**
          * As per ITU-T X.680 8824-1 8.6
@@ -427,6 +482,10 @@ inline fun <reified T : Asn1Element> T.assertTag(tag: Asn1Element.Tag): T {
 @Throws(Asn1TagMismatchException::class)
 inline fun <reified T : Asn1Element> T.assertTag(tagNumber: ULong): T = assertTag(tag withNumber tagNumber)
 
+/**
+ * ASN.1 NULL object as constant
+ */
+object Asn1Null : Asn1Primitive(Asn1Element.Tag.NULL, byteArrayOf())
 
 /**
  * ASN.1 structure. Contains no data itself, but holds zero or more [children]
@@ -667,7 +726,7 @@ sealed class Asn1Structure(
     }
 
 
-    override val length: Int by lazy { children.fold(0) { acc, child -> acc + child.overallLength } }
+    override val contentLength: Int by lazy { children.fold(0) { acc, child -> acc + child.overallLength } }
 
     override fun doEncode(sink: Sink) {
         children.let { childElems ->
@@ -816,6 +875,7 @@ class Asn1CustomStructure private constructor(
         shouldBeSorted: Boolean = false
     ) : this(children, tag.toULong(), tagClass, sortChildren, shouldBeSorted)
 
+    override fun toString() = "${tag.tagClass}" + super.toString()
 
     /**
      * Raw byte DER-encoded representation of this custom structure's children.
@@ -827,13 +887,11 @@ class Asn1CustomStructure private constructor(
         else null
     }
 
-    override fun toString() = "${tag.tagClass}" + super.toString()
-
     override fun prettyPrintHeader(indent: Int) =
         (" " * indent) + tag.tagClass +
                 " ${tag.tagValue}" +
                 (if (!tag.isConstructed) " PRIMITIVE" else "") +
-                " (=${tag.encodedTag.toHexString(HexFormat.UpperCase)}), length=${length}" +
+                " (=${tag.encodedTag.toHexString(HexFormat.UpperCase)}), length=${contentLength}" +
                 ", overallLength=${overallLength}" +
                 content?.let { " ${it.toHexString(HexFormat.UpperCase)}" }
 
@@ -869,6 +927,7 @@ class Asn1CustomStructure private constructor(
 class Asn1EncapsulatingOctetString(children: List<Asn1Element>) :
     Asn1Structure(Tag.OCTET_STRING, children, sortChildren = false, shouldBeSorted = false),
     Asn1OctetString {
+
     override val content: ByteArray by lazy {
         children.fold(byteArrayOf()) { acc, asn1Element -> acc + asn1Element.derEncoded }
     }
@@ -957,7 +1016,7 @@ open class Asn1Primitive(
         if (tag.isConstructed) throw IllegalArgumentException("A primitive cannot have a CONSTRUCTED tag")
     }
 
-    override val length: Int get() = content.size
+    override val contentLength: Int get() = content.size
     override fun doEncode(sink: Sink) {
         sink.write(tag.encodedTag)
         sink.write(encodedLength)
@@ -970,8 +1029,35 @@ open class Asn1Primitive(
 
     override fun prettyPrintHeader(indent: Int) = (" " * indent) + "Primitive" + super.prettyPrintHeader(indent)
 
-    override fun contentToString() = content.toHexString(HexFormat.UpperCase)
-    override fun prettyPrintContents(indent: Int) = contentToString()
+    override fun contentToString() = catchingUnwrapped {
+        when (tag) {
+            Tag.NULL -> ""
+            Tag.BOOL -> decodeToBoolean().toString()
+            Tag.INT -> decodeToInt().toString()
+            Tag.REAL -> decodeToFloat().toString()
+            Tag.OID -> ObjectIdentifier.decodeFromAsn1ContentBytes(content).let { oid ->
+                KnownOIDs[oid]?.let { "$it ($oid)" } ?: oid.toString()
+            }
+
+            Tag.ENUM -> decodeToEnumOrdinal().toString()
+            Tag.OCTET_STRING -> content.toHexString(HexFormat.UpperCase)
+            Tag.BIT_STRING -> content.toHexString(HexFormat.UpperCase)
+            Tag.STRING_UTF8 -> decodeToString()
+            Tag.STRING_UNIVERSAL -> decodeToString()
+            Tag.STRING_IA5 -> decodeToString()
+            Tag.STRING_BMP -> decodeToString()
+            Tag.STRING_T61 -> decodeToString()
+            Tag.STRING_PRINTABLE -> decodeToString()
+            Tag.STRING_NUMERIC -> decodeToString()
+            Tag.STRING_VISIBLE -> decodeToString()
+            Tag.TIME_GENERALIZED -> decodeToInstant().toString()
+            Tag.TIME_UTC -> decodeToInstant().toString()
+            else -> content.toHexString(HexFormat.UpperCase)
+        }
+    }.getOrElse { "Non-compliant content: 0x" + content.toHexString(HexFormat.UpperCase) }
+
+
+    override fun prettyPrintContents(indent: Int) = " " + contentToString()
 
 
     override fun hashCode() = 31 * super.hashCode() + content.contentHashCode()
@@ -997,7 +1083,7 @@ open class Asn1Primitive(
  * If you are expecting arbitrary bytes, [Asn1OctetString] is the correct type to look for.
  * (Your arbitrary bytes might inadvertently be valid ASN.1!)
  */
-interface Asn1OctetString {
+sealed interface Asn1OctetString {
 
     /**
      * Raw data contained in this ASN.1 primitive in its encoded form. Requires decoding to interpret it.
@@ -1010,6 +1096,7 @@ interface Asn1OctetString {
         /** Constructs a new ASN.1 OCTET STRING primitive containing these bytes */
         operator fun invoke(bytes: ByteArray) =
             Asn1PrimitiveOctetString(bytes)
+
         /** Constructs a new ASN.1 OCTET STRING primitive encapsulating these children */
         operator fun invoke(children: List<Asn1Element>) =
             Asn1EncapsulatingOctetString(children)